pre-release

+++ sberbank-dao/src/main/java/fssp38/sberbank/dao/CreateRequests.java
@@ -49,7 +49,8 @@ public class CreateRequests {
             } catch (Exception e) {
                 //в случае возникновения исключительной ситуации, удаляем сформированные файлы
                 e.printStackTrace();
-                service.deleteCreatedFiles();
+                if (service != null)
+                    service.deleteCreatedFiles();
             }
 
         }
